+++ app/src/main/res/layout/activity_search.xml
@@ -13,7 +13,7 @@
         android:layout_marginTop="10dp"
         android:layout_marginEnd="10dp"
         android:layout_marginBottom="10dp"
-        android:background="@color/gray"
+        android:background="@color/dark_gray"
         android:inputType="textCapWords"
         android:padding="5dp"
         app:layout_constraintStart_toStartOf="parent"
@@ -30,6 +30,16 @@
         app:layout_constraintTop_toBottomOf="@id/searchView"
         android:layout_height="wrap_content"
         android:layout_margin="16dp"/>
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recycler_view"
+        android:layout_width="0dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/chip_group"
+        app:layout_constraintBottom_toBottomOf="parent"
+        android:visibility="gone"
+        android:layout_height="0dp" />
 
     <androidx.gridlayout.widget.GridLayout
         android:id="@+id/gridView"
